+++ src/modules/start.ts
@@ -1,6 +1,6 @@
 import Client from "../client.js"
 import { Magic, Bit7 } from "../types/message-bytes.js"
-import { GamePlayerArray } from "../types/player-ds.js"
+import { GamePlayerArray } from "../types/list/player.js"
 import Player, { PlayerBase, SelfPlayer } from "../types/player/player.js"
 import World from "../types/world.js"
